fix: prevented circular import
