+++ src/conspiracies/__init__.py
@@ -1,5 +1,10 @@
 from .coref import CoreferenceComponent, create_coref_component  # noqa F401
 from .HeadWordExtractionComponent import HeadwordsExtractionComponent  # noqa F401
 from .HeadWordExtractionComponent import create_headwords_component  # noqa F401
-from .prompt_relation_extraction import PromptOutput, SpanTriplet  # noqa F401
+from .prompt_relation_extraction import (  # noqa F401
+    PromptOutput,
+    SpanTriplet,
+    docs_from_jsonl,
+    docs_to_jsonl,
+)
 from .wordpiece_length_normalization import wordpiece_length_normalization  # noqa F401
