+++ src/conspiracies/prompt_relation_extraction/__init__.py
@@ -1,1 +1,2 @@
-from .data_classes import PromptOutput, SpanTriplet
+from .data_classes import PromptOutput, SpanTriplet  # noqa F401
+from .utils import docs_from_jsonl, docs_to_jsonl  # noqa F401
